embeddings/requirements.txt
google-generativeai==0.5.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
gitpython==3.1.43
numpy==1.26.4